+++ NEWS.md
@@ -6,7 +6,7 @@
 
 ### Bug Fixes
 
-* Avoid spurious warnings in `robin_surv` during the hazard ratio estimation coming from the variance calculation: Now the variance is only calculated after the hazard ratio estimate is obtained.
+* There will be no more spurious warnings in `robin_surv` during the hazard ratio estimation coming from the variance calculation: Now the variance is only calculated after the hazard ratio estimate is obtained.
 * When adjusting for a single factor covariate, or when the covariate is strongly correlated with a strata variable in `robin_surv`, the function now works correctly and does not fail with a "singular design matrix" error any longer.
 
 # RobinCar2 0.2.0