+++ NEWS.md
@@ -8,6 +8,7 @@
 
 * There will be no more spurious warnings in `robin_surv` during the hazard ratio estimation coming from the variance calculation: Now the variance is only calculated after the hazard ratio estimate is obtained.
 * When adjusting for a single factor covariate, or when the covariate is strongly correlated with a strata variable in `robin_surv`, the function now works correctly and does not fail with a "singular design matrix" error any longer.
+* Fixed a bug in covariate-adjusted survival function estimation in `robin_surv` which could occur when there are character covariates with values only appearing in one treatment group, which could have led to incorrect results. 
 
 # RobinCar2 0.2.0